[Refactor] Use two kernels instead of CUDA cooperative kernel for batch/single decode (#72)
In our initial design we use CUDA cooperative kernels and grid
synchronization feature for cross threadblock reduction. Although it's
slightly faster for multi-head attention without grouping (GQA), we
found there are two issues with this implementation:
1. The kernel scheduling for cross-threadblock merging (the code after
`grid.sync`) is sub-optimal, the merging time would be the bottleneck
when the number of chunks to merge is huge (e.g. in GQA).
2. Not all hardware has grid synchronization feature (only NVIDIA & AMD
GPUs have such features AFAIK), which makes FlashInfer implementation
hard to generalize to other GPUs such as Metal.

In this PR we stop using CUDA cooperative kernels features and using the
combo two kernels (one for decode, another for merge) instead, it harms
performance a little bit for some shapes but I believe it's beneficial
for longer-term development and maintainance.