refactor: refactor `bmm_fp8.cuh` (#517)
This PR made refactors `bmm_fp8.cuh` to fix the following behaviors:
1. We should not instantiate templates in header files, if there are two
source files includes this header, there will be symbol conflict. So
it's preferable to instantiate templates in source files.
2. We don't want to depend on torch source code in the headers as torch
api's might subject to changes and other backends do not want to rely on
a specific version of torch. We should make the dependency of header
files as simple as possible (only cuda and libc++).

cc @zhyncs