misc: add bibtex reference (#712)
This pull request includes an update to the `README.md` file to add a
new citation section. The most important change is the addition of a
citation format for users who find FlashInfer helpful in their projects
or research.

Documentation update:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R144-R169):
Added a new "Citation" section with a BibTeX entry for citing the
FlashInfer paper.