misc: jit: Introduce JitSpec and Generate ninja file (#1065)
Part of AOT Refactor (#1064)

First, this PR introduces a `JitSpec` class for describing the build
instructions for a torch cpp extension. In future PRs, all modules will
return a `JitSpec` instead of a loaded torch ops. Then, AOT can build
all of these in one pass.

Second, instead of relying on torch, we generate the ninja files by
ourselves. This gives us more flexible control over the build process.
The major difference in the ninja file is that we prefix the target
files with the extension name (i.e., `$name/xxx.o` instead of `xxx.o`).
This allows us to use `subninja` to include all ops for the AOT build.