feat(aot): add nvshmem module for aot compilation
This change integrates the `gen_nvshmem_module` into the Ahead-of-Time (AOT) build process.

By including this module, users can now pre-compile the NVSHMEM-based communication kernels,
which is necessary for using this backend in environments where just-in-time compilation
is not available or desired.

Also install nvidia-nvshmem-cu12 in the CI container image.

Signed-off-by: Emilien Macchi <emacchi@redhat.com>