Fix install folder, and JIT-vs-AOT differences
This encompases a handful of fixes for build system errors.

1. `enable_aot` is true when aot_ops_package_dir is a symlink:
   https://github.com/flashinfer-ai/flashinfer/blob/main/setup.py#L27

   However, this is ALWAYS a symlink, the symlink is always created before
   the wheel is built, at
   https://github.com/flashinfer-ai/flashinfer/blob/main/custom_backend.py#L53
   so this change checks whether the AOT folder is empty or not instead.

2. root_is_pure handling changes the package's tag, but not the actual
   install behaviour. The code in
   https://github.com/flashinfer-ai/flashinfer/blob/main/setup.py#L72-L81
   does not successfully set the package as impure. This results in
   .so files being added to a purelib folder - see
   https://ci.tlcpack.ai/blue/rest/organizations/jenkins/pipelines/flashinfer-ci/branches/main/runs/374/nodes/10/steps/45/log/?start=0
   where files are being added to
   `[site-packages]/flashinfer_python-0.2.8.data/purelib/flashinfer/`
   whereas in older versions, they would go in
   `[site-packages]/flashinfer`. .so files in purelib is invalid, per
   PEP 427.

My changes a) cause the `if enable_aot` block to be skipped for jit
builds, b) cause the files to be installed to
`[site-packages]/flashinfer` as was the behaviour on older versions and
in accordance with PEP 427, but c) do not regress the platform tag
behaviour, where a JIT build produces
`flashinfer_python-0.2.8-py3-none-any.whl` and an AOT build produces
`flashinfer_python-0.2.8-cp39-abi3-linux_x86_64.whl`. The `-any.whl`
includes `Root-Is-Purelib: true` in `.dist-info/WHEEL`, the
`-linux_x86_64.whl` includes `Root-Is-Purelib: false`.