Fixing the JIT build command (#47)
There are 2 bugs in the exisitng jit compilaiton infrastructure

- The current infra does not compile kernels such as rope, page and norm
https://github.com/AMD-AIOSS/flashinfer/pull/46

- We recently discovered another bug where both the device and host code
was being compiled with the same set of flags. This was causing certain
parts of the code to build incorrectly as well as miss out on
optimization flags. This could lead to performance regressions in the
future.

This PR applies fixes to the second bug 

Tests:
- Installed both AOT and JIT and verified kernels
- Compared results with Nvidia tensors ensure correctness
- test_rope, test_norm and test_batch_decode_vllm

[Note: #45 was split into #46 and this PR]