fix: a workaround to make fp8 kv-cache work for prefill
This is for resolving an issue encountered while enabling fp8 kv-cache
support in the flashinfer backend:

https://github.com/vllm-project/vllm/pull/17005#issuecomment-2998112018

The root cause seems to be that we do not have native fp8 kv-cache
support for prefill. The failure that we hit, i.e.

```
static_assert(sizeof(ElementA) == 0, "No eligible GMMA operator for request configuration.");
```

is just a fact that our prefill kernels do not instantiate cute::GMMA::rs_op_selector
with the correct layout for fp8, which requires k-major for B matrix:

https://github.com/flashinfer-ai/flashinfer/blob/main/include/flashinfer/attention/hopper/kernel_traits.cuh#L78

Note that we cannot simply assign k-major when DTypeKV is of fp8.
There are more to fix to correctly support fp8 kv-cache in the kernel.

So, this comes to the workaround in this PR, where we convert k and v
to q_data_type if they are fp8 but q is not. We can do this from
vllm, but I think it seems to be better to put it in flashinfer, because
we do not require any changes to the customer code if we support
fp8 kv-cache for prefill in a better way.

Also please note that I am not 100% sure if this is an appropriate fix,
particularly I am not familiar with flashinfer's code base. Originally,
I was a bit worried about the impact to other kv-cache related things
such as _paged_kv_indptr and _kv_indptr_buf. It seems to be fine to
me after reading through the relevant code in prefill.py and
hopper/prefill_sm90.cuh.

Last note - eventually, I think we might need to support fp8 kv-cache
for prefill more appropriately.