Fix illegal mem access in routingIndicesWarpKernel
When all scores are -INF or NaNs, the resulting warpMaxExpertIdx[0] will
become 65535, and that causes illegal mem access at this line:

```
smemExpertTokenCountFull[tokenIdx][warpMaxExpertIdx[0] / ExpertsPerThread] =
            expertTokenCount;
```

To avoid this, add a check before accessing the shared memory. If
warpMaxExpertIdx[0] is out of range, we just set it to 0 as a fail-safe
mechanism.

Signed-off-by: Po-Han Huang <pohanh@nvidia.com>