Feature/load transposed 4x4 frag (#64)
The PR adds a new `load_fragment_4x4_half_registers` function to mma_hip.h. The function is meant to be used for a specific 4x4x4 thread layout in a wavefront that loads a "B" matrix from shared memory. The transposition allows us to change the layout of the fragment to what CDNA3 MFMA ops require for a fp32_16x16x16_fp16 MMA op.

The new fragment load transposes the 4 __half values stored in two 32b registers loaded by four adjacent threads in the inner 4x4 grid of the wavefront. The effect of the transformation is as follows:

```C++
    // Original data in registers for Threads 0-3 after fragment load
    // T0 : a b c d
    // T1 : e f g h
    // T2 : i j k l
    // T3 : m n o p
    //
    // After transposition:
    // T0 : a e i m
    // T1 : b f j n
    // T2 : c g k o
    // T3 : d h l p
````

---------

Co-authored-by: rtmadduri <rishi.madduri@amd.com>