Fix race condition when JitSpec loads the library
Currently, we only use FileLock when building the library, but does not
use FileLock when loading the library. This causes issue when process A
is loading the library while process B is building the same library and
removes the output .so file.

To fix this, use the same FileLock for building/loading the library in
JIT mode.

Signed-off-by: Po-Han Huang <pohanh@nvidia.com>