Support Kimi-K2 for TRT: templatize number of experts (#1696)
## Goal: Support Kimi-K2 Family of Models

The DeepSeek-v3 family of models use 256 experts, while the Kimi-K2
family of models use 384 experts. The number of experts is currently a
hardcoded constant in the DeepSeek routing kernels, and must remain a
compile-time constant in order to be used as the `__launchbounds__` for
some kernels.

  ### Proposed Solution: Template-configurable Number of Experts

In
[include/flashinfer/trtllm/fused_moe/RoutingKernel.h](https://github.com/flashinfer-ai/flashinfer/blob/main/include/flashinfer/trtllm/fused_moe/RoutingKernel.h),
change
  ```
  template <typename InputT_, typename OutputT_, bool UsePdl_>
  struct KernelParamsBase {
      ...
  }
  ```
  to
  ```
template <typename InputT_, typename OutputT_, bool UsePdl_, int
NumExperts_>
  struct KernelParamsBase {
      static constexpr bool NumExperts = NumExperts_;
      ...
  }
  ```
Then change
[csrc/trtllm_fused_moe_routing_deepseek.cu](https://github.com/flashinfer-ai/flashinfer/blob/main/csrc/trtllm_fused_moe_routing_deepseek.cu),
[csrc/trtllm_fused_moe_routing_renormalize.cu](https://github.com/flashinfer-ai/flashinfer/blob/main/csrc/trtllm_fused_moe_routing_renormalize.cu),
[csrc/trtllm_fused_moe_routing_llama4.cu](https://github.com/flashinfer-ai/flashinfer/blob/main/csrc/trtllm_fused_moe_routing_llama4.cu),
to use `KernelParams::NumExperts` as the number of experts (might need
some boilerplate to "pass" it from `KernelParamsBase` to `KernelParams`
as well).


  ### Benefits
  - Easy to support different numbers of experts in the future.
  - Zero runtime overhead.

  ### Costs
- (NEW) The numbers of experts for Llama4 and `renormalize` kernels now
need to be known at compile time, where previously only the *max* number
of experts was known at compile time. This means that `flashinfer` users
passing custom numbers of experts to these kernels will have their code
broken if we don't include that number of experts in the `run` dispatch
logic. If this is a concern we *might* be able to keep the number of
experts defined at compile time in the deepseek kernel and at runtime in
the Llama4 and renormalize kernels, but this would come at the cost of
some complexity.

  ### Alternatives Considered

1. New file called `csrc/trtllm_fused_moe_routing_kimi.cu` that is an
exact copy of
[csrc/trtllm_fused_moe_routing_deepseek.cu](https://github.com/flashinfer-ai/flashinfer/blob/main/csrc/trtllm_fused_moe_routing_deepseek.cu)
except it hardcodes 384 experts. This makes maintenance of both files
more difficult, and doesn't make it easy to support different numbers of
experts in the future.

2. Make
[csrc/trtllm_fused_moe_routing_deepseek.cu](https://github.com/flashinfer-ai/flashinfer/blob/main/csrc/trtllm_fused_moe_routing_deepseek.cu)
accept a compiler flag defining the number of experts (`-DNUM_EXPERTS
256` and `-DNUM_EXPERTS 384`). This moves complexity to the build stage,
which doesn't seem like the right place to manage it.