fix: missing header include in decode kernel jit binding
This commit adds the missing include of the header `decode.cuh`
in the JIT binding `csrc/batch_decode.cu`, without this fix,
there will be linking error when linking `batch_decode.o`,
`batch_decode_kernel.o` and `batch_decode_jit_binding.o`.

The reason is that, `csrc/batch_decode.cu` relies on `scheduler.cuh`
for decoding attention planning. `scheduler.cuh` declares the decode
kernel of `BatchDecodeWithPagedKVCacheKernel`, which is implemented
in `decode.cuh`, but does not include `decode.cuh`.  Therefore,
if `csrc/batch_decode.cu` does not include `decode.cuh`, the decode
kernel will be a symbol that is declared but not implemented.