Add bit-width specific masks (#7)
The mask values for the __shfl_xor_sync used in the original CUDA version of the frag_layout_swizzle.cuh was designed for 32-thread warps. The design made the header incompatible with CDNA3 that has 64-thread warps. The PR adds a platform-specific
WARP_FULL_MASK constant to support both 32 thread and 64 thread warps.