bugfix: show-config command (#1846)
<!-- .github/pull_request_template.md -->

## 📌 Description

I believe the show-config command line option was broken by #1794,
because it change the return type of `get_artifacts_status` from a list
of size 3 tuples to a tuple of size 2 tuples. This code in `__main__.py`
expects the former:

```
    status = get_artifacts_status()
    num_downloaded = sum(1 for _, _, exists in status if exists)
    total_cubins = len(status)
```

This PR has a fix and a test to protect against future breaks.

## 🔍 Related Issues

<!-- Link any related issues here -->

## 🚀 Pull Request Checklist

Thank you for contributing to FlashInfer! Before we review your pull
request, please make sure the following items are complete.

### ✅ Pre-commit Checks

- [x] I have installed `pre-commit` by running `pip install pre-commit`
(or used your preferred method).
- [x] I have installed the hooks with `pre-commit install`.
- [x] I have run the hooks manually with `pre-commit run --all-files`
and fixed any reported issues.

> If you are unsure about how to set up `pre-commit`, see [the
pre-commit documentation](https://pre-commit.com/).

## 🧪 Tests

- [x] Tests have been added or updated as needed.
- [x] All tests are passing (`unittest`, etc.).

## Reviewer Notes

Is this the right place to add CLI tests? Note that I've added a
`tests/cli` subdirectory.